+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/firFunctionCall.kt
@@ -6,11 +6,13 @@ import org.jetbrains.kotlin.fir.expressions.FirFunctionCall
 import org.jetbrains.kotlin.fir.expressions.buildResolvedArgumentList
 import org.jetbrains.kotlin.fir.expressions.builder.buildFunctionCall
 import org.jetbrains.kotlin.fir.references.builder.buildResolvedNamedReference
+import org.jetbrains.kotlin.fir.symbols.impl.FirClassLikeSymbol
 
-public inline val FirSession.nothingFirFunctionCall: FirFunctionCall
+context(session: FirSession)
+public inline val nothingFirFunctionCall: FirFunctionCall
     get() = buildFunctionCall {
-        val nothingSymbol =
-            builtinTypes.nothingType.toClassLikeSymbol(this@nothingFirFunctionCall)!!
+        val nothingSymbol: FirClassLikeSymbol<*> =
+            session.builtinTypes.nothingType.toClassLikeSymbol(session)!!
         argumentList = buildResolvedArgumentList(null, LinkedHashMap())
         calleeReference = buildResolvedNamedReference {
             this.name = nothingSymbol.name