+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/declarationIrBuilder.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package com.javiersc.kotlin.compiler.extensions.ir
 
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -5,9 +7,10 @@ import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
 import org.jetbrains.kotlin.ir.declarations.IrSymbolOwner
 import org.jetbrains.kotlin.ir.symbols.IrSymbol
 
-public inline fun IrPluginContext.declarationIrBuilder(symbol: IrSymbol): DeclarationIrBuilder =
-    DeclarationIrBuilder(this, symbol)
+context(context: IrPluginContext)
+public inline fun DeclarationIrBuilder(symbol: IrSymbol): DeclarationIrBuilder =
+    DeclarationIrBuilder(generatorContext = context, symbol = symbol)
 
-public inline fun IrPluginContext.declarationIrBuilder(
-    element: IrSymbolOwner
-): DeclarationIrBuilder = DeclarationIrBuilder(this, element.symbol)
+context(context: IrPluginContext)
+public inline fun DeclarationIrBuilder(element: IrSymbolOwner): DeclarationIrBuilder =
+    DeclarationIrBuilder(generatorContext = context, symbol = element.symbol)