+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irBlockBody.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package com.javiersc.kotlin.compiler.extensions.ir
 
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -5,20 +7,25 @@ import org.jetbrains.kotlin.ir.IrStatement
 import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
 import org.jetbrains.kotlin.ir.expressions.IrBlockBody
 
-public inline fun IrPluginContext.createIrBlockBody(
+context(context: IrPluginContext)
+public inline fun createIrBlockBody(
     startOffset: Int = UNDEFINED_OFFSET,
     endOffset: Int = UNDEFINED_OFFSET,
-): IrBlockBody = irFactory.createBlockBody(startOffset, endOffset)
+): IrBlockBody = context.irFactory.createBlockBody(startOffset, endOffset)
 
-public inline fun IrPluginContext.createIrBlockBody(
+context(context: IrPluginContext)
+public inline fun createIrBlockBody(
     startOffset: Int = UNDEFINED_OFFSET,
     endOffset: Int = UNDEFINED_OFFSET,
     initializer: IrBlockBody.() -> Unit,
-): IrBlockBody = irFactory.createBlockBody(startOffset, endOffset).apply(initializer)
+): IrBlockBody = context.irFactory.createBlockBody(startOffset, endOffset).apply(initializer)
 
-public inline fun IrPluginContext.createIrBlockBody(
+context(context: IrPluginContext)
+public inline fun createIrBlockBody(
     startOffset: Int = UNDEFINED_OFFSET,
     endOffset: Int = UNDEFINED_OFFSET,
     statements: List<IrStatement>,
 ): IrBlockBody =
-    irFactory.createBlockBody(startOffset, endOffset).apply { this.statements.addAll(statements) }
+    context.irFactory.createBlockBody(startOffset, endOffset).apply {
+        this.statements.addAll(statements)
+    }