+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irExpressionBody.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package com.javiersc.kotlin.compiler.extensions.ir
 
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -5,12 +7,13 @@ import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
 import org.jetbrains.kotlin.ir.expressions.IrExpression
 import org.jetbrains.kotlin.ir.expressions.IrExpressionBody
 
-public inline fun IrPluginContext.createExpressionBody(
+context(context: IrPluginContext)
+public inline fun createExpressionBody(
     startOffset: Int = UNDEFINED_OFFSET,
     endOffset: Int = UNDEFINED_OFFSET,
     expression: IrExpression,
 ): IrExpressionBody =
-    irFactory.createExpressionBody(
+    context.irFactory.createExpressionBody(
         startOffset = startOffset,
         endOffset = endOffset,
         expression = expression,