+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `org.jetbrains.kotlin:kotlin-gradle-plugin -> 2.3.20`
 - `gradle -> 9.4.0`
 
 ## [0.12.0+2.3.10] - 2026-02-06